+++ src/k8s-extension/azext_k8s_extension/_params.py
@@ -29,9 +29,12 @@ def load_arguments(self, _):
                    options_list=['--cluster-name', '-c'],
                    help='Name of the Kubernetes cluster')
         c.argument('cluster_type',
-                   arg_type=get_enum_type(['connectedClusters', 'managedClusters', 'appliances']),
+                   arg_type=get_enum_type(['connectedClusters', 'managedClusters', 'appliances', 'provisionedClusters']),
                    options_list=['--cluster-type', '-t'],
-                   help='Specify Arc clusters or AKS managed clusters or Arc appliances.')
+                   help='Specify Arc clusters or AKS managed clusters or Arc appliances or provisionedClusters.')
+        c.argument('cluster_resource_provider',
+                   options_list=['--cluster-resource-provider', '--cluster-rp'],
+                   help='Cluster Resource Provider name for this clusterType (Required for provisionedClusters)')
         c.argument('scope',
                    arg_type=get_enum_type(['cluster', 'namespace']),
                    help='Specify the extension scope.')
@@ -84,3 +87,33 @@ def load_arguments(self, _):
                    help='Ignore confirmation prompts')
         c.argument('force',
                    help='Specify whether to force delete the extension from the cluster.')
+
+    with self.argument_context(f"{consts.EXTENSION_NAME} extension-types list") as c:
+        c.argument('cluster_name',
+                   options_list=['--cluster-name', '-c'],
+                   help='Name of the Kubernetes cluster')
+        c.argument('cluster_type',
+                   arg_type=get_enum_type(['connectedClusters', 'managedClusters', 'appliances']),
+                   options_list=['--cluster-type', '-t'],
+                   help='Specify Arc clusters or AKS managed clusters or Arc appliances.')
+
+    with self.argument_context(f"{consts.EXTENSION_NAME} extension-types list-by-location") as c:
+        c.argument('location',
+                   validator=get_default_location_from_resource_group)
+
+    with self.argument_context(f"{consts.EXTENSION_NAME} extension-types show") as c:
+        c.argument('extension_type',
+                   help='Name of the extension type.')
+        c.argument('cluster_name',
+                   options_list=['--cluster-name', '-c'],
+                   help='Name of the Kubernetes cluster')
+        c.argument('cluster_type',
+                   arg_type=get_enum_type(['connectedClusters', 'managedClusters', 'appliances']),
+                   options_list=['--cluster-type', '-t'],
+                   help='Specify Arc clusters or AKS managed clusters or Arc appliances.')
+
+    with self.argument_context(f"{consts.EXTENSION_NAME} extension-types list-versions") as c:
+        c.argument('extension_type',
+                   help='Name of the extension type.')
+        c.argument('location',
+                   validator=get_default_location_from_resource_group)